requirements-dev.txt
# Dépendances de développement pour pynetro

# Tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Client HTTP pour les tests d'intégration
aiohttp>=3.8.0

# Tests en parallèle (optionnel)
pytest-xdist>=3.0.0

# Linting (déjà inclus dans pyproject.toml mais peut être installé séparément)
ruff>=0.1.0

# Chargement automatique des variables d'environnement
python-dotenv>=1.0.0